+++ SzamlazzHuSDK/Requests/CreateInvoiceResponse.cs
@@ -6,9 +6,9 @@ namespace SzamlazzHu
         public int ErrorCode { get; set; }
         public string ErrorMessage { get; set; }
         public string InvoiceNumber { get; set; }
-        public float NetPrice { get; set; }
-        public float GrossPrice { get; set; }
-        public float Receivable { get; set; }
+        public decimal NetPrice { get; set; }
+        public decimal GrossPrice { get; set; }
+        public decimal Receivable { get; set; }
         public string CustomerAccountUrl { get; set; }
         public byte[] InvoicePdf { get; set; }
     }
